Déplacement du enpoint /api/getNbPropositionsParProposeur dans un contrôleur dédié et fix requête pour ne pas prendres les semaines sans PS ou sans Film
